added usage monitoring
enabling of usage monitoring is controlled via the .env file by specifying ENFORCE_USER_STORAGE_SIZE=True
setting the default usage amount is through a similar setting in the .env file USER_STORAGE_LIMIT_BYTES=50000000 for a usage limite of 50bm.  Storage is set by bytes
setting of user storage is set when users access the dashboard page.  This is not the best way to enable this but proper setting would require a callback on google auth library which we currently do not have under version control
users total usage is displayed on the dashboard,
usage is set during file upload, if a file upload is greater than the allotted storage for the user, a warning is set and user cannot proceed
when projects or studies are deleted, the contents are not wiped from the system, only the active column of the tables are flipped to false.  To keep in line with this decision, storage amount for accounts are only decremented by the amount of storage a project or study used and the files are not deleted from the system